move context to Context directory
