+++ Contexts/RecipleaseContext.cs
@@ -1,0 +1,21 @@
+using Microsoft.EntityFrameworkCore;
+namespace Reciplease;
+
+public class RecipleaseContext : DbContext 
+{
+    public DbSet<Recipe> Recipes { get; set; }
+    public DbSet<Ingredient> Ingredients { get; set; }
+    public DbSet<RecipeIngredient> RecipeIngredients { get; set; }
+
+    public string DbPath { get; }
+
+    public RecipleaseContext() 
+    {
+        var folder = Environment.SpecialFolder.LocalApplicationData;
+        var path = Environment.GetFolderPath(folder);
+        DbPath = System.IO.Path.Join(path, "reciplease.db");     
+    }
+
+    protected override void OnConfiguring(DbContextOptionsBuilder options)
+        => options.UseSqlite($"Data Source={DbPath}");
+}
